+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # CHANGELOG
 
+## 0.1.1 (2026-03-10)
+
+### Added
+
+ - None
+
+### Changed
+
+- Update simulated time calculator input field to allow delete all value
+- Update simulated time calculator fixed value to change from `1, 6, 12, ..., 60` to `1, 5, 10, ..., 60`
+
+### Removed
+
+- None
+
 ## 0.1.0 (2026-03-07)
 
 ### Added